+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Add support for global rules-exclude
 * Add `Halstead Volume` metric.
 
 ## 4.2.1